Fixed RightClick event for egg drop

+++ java/com/branders/spawnermod/config/SpawnConfig.java
@@ -4,7 +4,7 @@ import net.minecraftforge.common.ForgeConfigSpec;
 
 public class SpawnConfig 
 {
-	public static ForgeConfigSpec.IntValue monster_egg_drop_chance;
+	public static ForgeConfigSpec.DoubleValue monster_egg_drop_chance;
 	
 	public static void init(ForgeConfigSpec.Builder common)
 	{
@@ -12,6 +12,6 @@ public class SpawnConfig
 		
 		monster_egg_drop_chance = common
 				.comment("Drop chance for monster eggs. Value in percentage (%) ranging from 0 - 100. Default is 4")
-				.defineInRange("spawnconfig.monster_egg_drop_chance", 4, 0, 100);
+				.defineInRange("spawnconfig.monster_egg_drop_chance", 4D, 0D, 100D);
 	}
 }
